requirements.txt
notion-client==2.2.1
requests==2.32.3
python-dotenv==1.0.1
mistralai==0.0.7 